refactor(types): Move Inquiry type to interfaces directory
- Inquiry 타입을 types/interfaces 디렉토리로 이동하여
  코드 구조를 개선했습니다.
