+++ src/types/interfaces/Inquiry.ts
@@ -1,4 +1,3 @@
-// src/types/Inquiry.ts
 export interface Inquiry {
   subject: string;
   category: string;
